+++ R/utils-manifest.R
@@ -9,22 +9,18 @@ retrieve_manifest <- function ( ) {
     `compare_cells`       = readr::col_logical(),
     `md5`                 = readr::col_character()
   )
-  d <- readr::read_csv(path, col_types = col_types)
-
-  d
+  readr::read_csv(path, col_types = col_types)
 }
 # retrieve_manifest()
 
 retrieve_file_expected <- function (subdataverse, file_name) {
-  path <-
-    file.path(subdataverse, file_name) %>%
-    system.file(package = "dataverse")
+  path <- system.file(file.path(subdataverse, file_name), package = "dataverse")
 
   if (!file.exists(path)) {
     stop(
       "The testing file `",
       file_name,
-      "` in the (sub)datavserse `",
+      "` in the (sub)dataverse `",
       subdataverse,
       "` is not found.  Please verify that the manifest is synced with the collection of test files. "
     )
@@ -32,8 +28,17 @@ retrieve_file_expected <- function (subdataverse, file_name) {
 
   readr::read_file(path)
 }
-# retrieve_file_expected("rosters", "roster-bulls-1996.csv")
+# retrieve_file_expected("dataset-basketball/original", "roster-bulls-1996.csv")
 
+retrieve_info_dataverse <- function (path) {
+  yaml::read_yaml(system.file(path, package = "dataverse"))
+}
+# retrieve_info_dataverse("expected-dataverse.yml")
+
+retrieve_info_dataset <- function (path) {
+  yaml::read_yaml(system.file(path, package = "dataverse"))
+}
+# retrieve_info_dataset("dataset-basketball/expected-metadata.yml")
 
 # compare_data_frame <- function (d_actual, d_expected) {
 #