+++ R/utils.R
@@ -81,6 +81,39 @@ get_fileid.dataverse_file <- function(x, ...) {
     x[["dataFile"]][["id"]]
 }
 
+
+#' Identify if file is an ingested file
+#'
+#' @param fileid A numeric fileid or file-specific DOI
+#' @template envvars
+#'
+# @examples
+# # https://demo.dataverse.org/file.xhtml?persistentId=doi:10.70122/FK2/X5MUPQ/T0KKUZ
+# # nlsw88.tab
+# is_ingested(fileid = "doi:10.70122/FK2/X5MUPQ/T0KKUZ",
+#             server = "demo.dataverse.org")
+#
+# # nlsw88_rds-export.rds
+# is_ingested(fileid = "doi:10.70122/FK2/PPIAXE/SUCFNI",
+#             server = "demo.dataverse.org")
+#
+is_ingested <-
+  function(
+    fileid,
+    key     = Sys.getenv("DATAVERSE_KEY"),
+    server  = Sys.getenv("DATAVERSE_SERVER")
+  ) {
+    ping_metadata <- tryCatch(
+      {
+        get_file_metadata(fileid, key = key, server = server)
+      },
+      error = function(e) e
+    )
+    !inherits(ping_metadata, "error") # if error, not ingested
+}
+
+
+
 # other functions
 prepend_doi <- function(dataset) {
     if (grepl("^hdl", dataset)) {
@@ -100,7 +133,6 @@ prepend_doi <- function(dataset) {
     dataset
 }
 
-#' @import httr
 api_url <- function(server = Sys.getenv("DATAVERSE_SERVER"), prefix = "api/") {
     if (is.null(server) || server == "") {
         stop("'server' is missing with no default set in DATAVERSE_SERVER environment variable.")
@@ -143,3 +175,4 @@ parse_dataset <- function(out) {
     out$files <- cbind(out$files, file_df)
     structure(out, class = "dataverse_dataset")
 }
+