+++ man-roxygen/ds.R
@@ -1,1 +1,3 @@
-#' @param dataset An integer specifying a dataset identification number or an object of class \dQuote{dataverse_dataset}. The identification number is the dataset's persistent identification number (not the integer specifying a specific version of the dataset, such as returned by \code{\link{dataset_versions}}).
+#' @param dataset A character specifying a persistent identification ID for a dataset,
+#'  for example `"doi:10.70122/FK2/HXJVJU"`. Alternatively, an object of class
+#'  \dQuote{dataverse_dataset} obtained by `dataverse_contents()`.