+++ man-roxygen/envvars.R
@@ -1,2 +1,8 @@
-#' @param key A character string specifying a Dataverse server API key. If one is not specified, functions calling authenticated API endpoints will fail. Keys can be specified atomically or globally using \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.
-#' @param server A character string specifying a Dataverse server. There are multiple Dataverse installations, but the defaults is to use the Harvard Dataverse. This can be modified atomically or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.
+#' @param key A character string specifying a Dataverse server API key. If one
+#'   is not specified, functions calling authenticated API endpoints will fail.
+#'   Keys can be specified atomically or globally using
+#'   \code{Sys.setenv("DATAVERSE_KEY" = "examplekey")}.
+#' @param server A character string specifying a Dataverse server. There are
+#'   multiple Dataverse installations, but the defaults is to use the Harvard
+#'   Dataverse (`server = "dataverse.harvard.edu"`). This can be modified atomically
+#'   or globally using \code{Sys.setenv("DATAVERSE_SERVER" = "dataverse.example.com")}.