+++ NEWS.md
@@ -4,8 +4,9 @@
 * Update maintainer to Will Beasley (wibeasley@hotmail.com) (#38)
 * More robust file retrieval (#39 @kuriwaki)
 * Tests use https://demo.dataverse.org/dataverse/dataverse-client-r/. (#40)
-* Fixes most get_file errors by removing query argument (#33 kuriwaki)
+* Fixes most get_file errors by removing query argument (#33 @kuriwaki)
 * Fix getting multiple files by id in `get_file()` (#47 @adam3smith)
+* Temporary files created by `get_file()` are automatically deleted.
 
 # CHANGES TO dataverse 0.2.1